requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
gunicorn
gevent
whitenoise
psycopg2-binary